DAAS-194 binary gets deleted when sample is deleted + other stuff (#63)
- When uploaded a sample, the code used to delete and recreate the
associated Sample object, to maintain the 1-to-1 relationship even if
the sample already existed. This caused the app to stop working when we
added that Sample's delete method also deletes the binary from
SeaWeedFS. We have changed the schema to 1-to-many, and show the user
the last object.
- Change verbosity level to DEBUG
- Create a favicon path so that we stop getting the 404 stacktrace on
the logs